feat: 添加赛季管理菜单并优化页面布局
1. 新增赛季管理菜单，扩展主导航功能
2. 实现固定顶部菜单布局，菜单始终在浏览器顶部
3. 优化页面内容自适应，支持完全显示所有内容
4. 修复人均战功增量（加成后）计算逻辑的类型转换问题
5. 确保页面响应式设计，适应不同浏览器尺寸
6. 移除内容截断限制，允许页面内容完全展示

+++ backend/src/main/java/com/tripmaster/backend/attendance/AttendanceController.java
@@ -348,10 +348,10 @@ public class AttendanceController {
                 int memberCount = stat.getMemberCount();
                 if (memberCount >= 40 && memberCount <= 45) {
                     // 小组人数40-45，加成1.03
-                    averageMeritIncreaseBonus = Math.round(averageMeritIncrease * 1.03);
+                    averageMeritIncreaseBonus = Math.round((double) averageMeritIncrease * 1.03);
                 } else if (memberCount >= 46 && memberCount <= 50) {
                     // 小组人数46-50，加成1.05
-                    averageMeritIncreaseBonus = Math.round(averageMeritIncrease * 1.05);
+                    averageMeritIncreaseBonus = Math.round((double) averageMeritIncrease * 1.05);
                 }
                 
                 stat.setAverageMeritIncreaseBonus(averageMeritIncreaseBonus);
